% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oncoAgent_version}
\alias{oncoAgent_version}
\title{Utility Functions}
\usage{
oncoAgent_version()
}
\value{
Version string
}
\description{
Package-level utility functions and configuration.

Returns the current version of the oncoAgent package.
}
