% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchestration.R
\name{pipeline_status}
\alias{pipeline_status}
\title{Pipeline Status}
\usage{
pipeline_status(result)
}
\arguments{
\item{result}{Output from run_oncopipeline()}
}
\value{
Formatted status string
}
\description{
Returns current status of a pipeline run
}
