% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{recall_of_known_targets}
\alias{recall_of_known_targets}
\title{Recall of Known Targets}
\usage{
recall_of_known_targets(predicted, known_targets)
}
\arguments{
\item{predicted}{Character vector of predicted genes}

\item{known_targets}{Character vector of known gold standard genes}
}
\value{
Recall value (0-1)
}
\description{
Computes recall (sensitivity) for recovering known targets.
}
