% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_oncoagent_config}
\alias{set_oncoagent_config}
\title{Set oncoAgent Configuration}
\usage{
set_oncoagent_config(
  cache_dir = NULL,
  n_cores = 1,
  verbose = TRUE,
  log_level = "INFO"
)
}
\arguments{
\item{cache_dir}{Default cache directory}

\item{n_cores}{Number of parallel cores to use}

\item{verbose}{Logical, enable verbose logging}

\item{log_level}{Logging level: "INFO", "DEBUG", "WARN", "ERROR"}
}
\value{
Invisibly returns the configuration list
}
\description{
Sets global configuration options for the oncoAgent package.
}
